+++ client/src/client.js
@@ -1,0 +1,3 @@
+import Gallery from './index.jsx';
+
+window.Galley = Gallery;
