+++ client/src/index.jsx
@@ -4,11 +4,12 @@ import Lightbox from 'react-image-lightbox';
 import { Button, Icon } from 'react-materialize';
 import Modal from 'react-modal';
 import axios from 'axios';
-import OpeningPageGalleryView from './openingGrid.jsx';
-import FullGalleryOpenGrid from './fullGalleryOpenGrid.jsx';
-import Header from './Header.jsx';
-import Social from './social.jsx';
-import '../dist/style.css';
+import OpeningPageGalleryView from './components/openingGrid.jsx';
+import FullGalleryOpenGrid from './components/fullGalleryOpenGrid.jsx';
+import Header from './components/Header.jsx';
+import Social from './components/social.jsx';
+//import '../dist/style.css';
+//import window from 'global/window';
 
 
 class ApateezGallery extends React.Component {
@@ -32,8 +33,12 @@ class ApateezGallery extends React.Component {
     Modal.setAppElement(document.getElementById('gallery'));
   }
   componentDidMount() {
-    const id = window.location.href.split('restaurants/')[1];
-    this.getRequestWithId(id);
+    // const id = window.location.href.split('restaurants/')[1];
+    console.log('component did mount');
+    const id = this.props.place_id;
+    if (typeof(window) !== 'undefined') {
+      this.getRequestWithId(id);
+    }
   }
 
   // getRequestWithId(id) {
@@ -49,9 +54,11 @@ class ApateezGallery extends React.Component {
   getRequestWithId(id) {
     //const appContext = this;
     // `${BASE_URL}/api/restaurants/${id}/gallery`
+    console.log('component did mount');
     axios.get(`http://localhost:3002/api/restaurants/${id}/gallery`)
       .then(({ data }) => {
         this.setState({ images: data.photoArray, restaurantName: data.restaurantName, place_id: data.place_id });
+        console.log(this.state);
       })
       .catch(err => console.error(err));
   }
