+++ client/src/server.js
@@ -1,0 +1,3 @@
+import App from './index.jsx';
+
+export default App;
