+++ package.json
@@ -49,6 +49,7 @@
     "express": "^4.15.0",
     "faker": "^4.1.0",
     "fs": "0.0.1-security",
+    "global": "^4.3.2",
     "http": "0.0.0",
     "jquery": "^3.1.1",
     "json-schema-faker": "^0.5.0-rc15",
